refactor: calculator 분리 및 allClear 메서드 수정
